✅ Ticket list, My ticket list 공통 컴포넌트, 로직 모듈화 & 공통 utils 디렉토리 생성
