+++ app/my-page/my-ticket-list/page.tsx
@@ -2,7 +2,7 @@ import { Metadata } from "next";
 import MyTicktListPage from "app/my-page/my-ticket-list/my-ticket-list-page";
 
 export const metadata: Metadata = {
-  title: "My Ticket List",
+  title: "My Tickets",
 };
 
 export default function Page() {
