+++ app/ticket-list/ticket-list-page.tsx
@@ -7,11 +7,11 @@ import { useForm } from "react-hook-form";
 import fetchMovieReviews, {
   MovieReview,
 } from "api/movie-reviews/fetchMovieReviews";
-import { firebaseErrorHandler } from "app/my-page/utils/firebase-error";
+import { firebaseErrorHandler } from "app/utils/firebase-error";
 import { addNewReviewAlertHandler } from "store/newReviewAlertSlice";
-import TicketList from "app/ticket-list/ticket-list";
-import { IoSearchOutline } from "react-icons/io5";
-import useReviewSearch from "app/ticket-list/utils/useReviewSearch";
+import useReviewSearch from "app/utils/useReviewSearch";
+import ReviewTicket from "app/ui/reviewTicketList/review-ticket";
+import ReviewSearchInputregister from "app/ui/reviewTicketList/review-search-Input";
 
 export default function TicketListPage() {
   const newReviewAlertState = useAppSelector(
@@ -61,7 +61,7 @@ export default function TicketListPage() {
   }, [searchTerm, searchReviewsHandler]);
 
   return (
-    <div className="w-full px-8 md:mt-6">
+    <main className="w-full px-8 md:mt-6">
       <section className="flex w-full flex-col items-center pb-6 md:flex-row">
         <div className="mb-4 mt-8 flex w-full items-center justify-between md:mb-0 md:mt-0 md:justify-normal">
           <h1 className="text-2xl font-bold">ALL TICKETS</h1>
@@ -70,37 +70,23 @@ export default function TicketListPage() {
           </span>
         </div>
         {/* 티켓 검색 */}
-        <div className="relative flex h-10 w-full items-center justify-end">
-          <label htmlFor="review-search" className="sr-only">
-            리뷰 검색
-          </label>
-          <input
-            {...register("search")}
-            id="review-search"
-            type="search"
-            placeholder="티켓 검색"
-            className="h-full w-full rounded-full border-2 border-black pl-4 pr-10 text-sm opacity-100 md:w-64"
-          />
-          <div
-            className={`absolute right-0 top-0 flex h-full w-10 cursor-pointer items-center justify-center rounded-full`}
-          >
-            <IoSearchOutline size={20} color="black" />
-          </div>
-        </div>
+        <ReviewSearchInputregister
+          label="티켓 검색"
+          register={register}
+          placeholder="티켓 검색"
+        />
       </section>
       {/* 티켓 목록 */}
       {filteredUserReviews ? (
-        <main className="grid grid-cols-1 gap-2 pb-8 md:grid-cols-3 md:gap-4 lg:grid-cols-4 xl:grid-cols-5">
-          <TicketList
-            reviews={filteredUserReviews}
-            onReviewUpdated={fetchReviews}
-          />
-        </main>
+        <ReviewTicket
+          reviews={filteredUserReviews}
+          onReviewUpdated={fetchReviews}
+        />
       ) : (
         <div className="p-8 text-center text-sm font-bold text-gray-300 md:pt-60 md:text-xl">
           리뷰가 없습니다
         </div>
       )}
-    </div>
+    </main>
   );
 }
